+++ Questionable/Controller/Steps/Interactions/Emote.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using Questionable.Controller.Steps.Common;
+using Questionable.Controller.Steps.Shared;
 using Questionable.Domain;
 using Questionable.Functions;
 using Questionable.Model.Questing;
@@ -8,7 +9,7 @@ namespace Questionable.Controller.Steps.Interactions;
 
 internal static class Emote
 {
-    internal sealed class Factory : ITaskFactory
+    internal sealed class Factory(Configuration configuration) : ITaskFactory
     {
         public IEnumerable<ITask> CreateAllTasks(Quest quest, QuestSequence sequence, QuestStep step)
         {
@@ -16,24 +17,28 @@ internal static class Emote
                 or EInteractionType.SinglePlayerDuty)
             {
                 if (step.Emote == null)
-                    return [];
+                    yield break;
+                if (step.InteractionType is EInteractionType.CompleteQuest)
+                {
+                    yield return new LogQuestCompletion.Task(quest);
+                    if (configuration.Advanced.PreventQuestCompletion)
+                    {
+                        if (configuration.Advanced.AbandonQuestBeforeCompletion)
+                            yield return new AbandonQuest.Task(quest);
+                        yield break;
+                    }
+                }
             }
             else if (step.InteractionType != EInteractionType.Emote)
-                return [];
+                yield break;
             if (!step.Emote.HasValue)
                 throw new ArgumentNullException(nameof(step.Emote));
 
-            Mount.UnmountTask unmount = new();
+            yield return new Mount.UnmountTask();
             if (step.DataId != null)
-            {
-                UseOnObject task = new(step.Emote.Value, step.DataId.Value);
-                return [unmount, task];
-            }
+                yield return new UseOnObject(step.Emote.Value, step.DataId.Value);
             else
-            {
-                UseOnSelf task = new(step.Emote.Value);
-                return [unmount, task];
-            }
+                yield return new UseOnSelf(step.Emote.Value);
         }
     }